--- a/README.md
+++ b/README.md
@@ -1,7 +1,7 @@
 # 🚀 Free Deep Research System
 
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](LICENSE)
-[![Version](https://img.shields.io/badge/version-4.9.0-blue.svg)](CHANGELOG.md)
+[![Version](https://img.shields.io/badge/version-5.0.0-blue.svg)](CHANGELOG.md)
 [![Production Ready](https://img.shields.io/badge/production-ready-green.svg)](PRODUCTION_DEPLOYMENT_GUIDE.md)
 [![Enterprise Grade](https://img.shields.io/badge/enterprise-grade-gold.svg)](PROJECT_COMPLETION_SUMMARY.md)
 [![MLOps](https://img.shields.io/badge/MLOps-enabled-purple.svg)](infrastructure/kubernetes/README-PHASE-4.6.md)
@@ -10,17 +10,30 @@
 [![Analytics](https://img.shields.io/badge/analytics-real--time-brightgreen.svg)](infrastructure/kubernetes/README-PHASE-4.7.md)
 [![Kubernetes](https://img.shields.io/badge/kubernetes-native-blue.svg)](infrastructure/kubernetes/)
 [![AI Powered](https://img.shields.io/badge/AI-powered-ff69b4.svg)](docs/user-guides/bmad-agents.md)
+[![RAG Enabled](https://img.shields.io/badge/RAG-enabled-success.svg)](infrastructure/kubernetes/phase-5.0/)
+[![Vector DB](https://img.shields.io/badge/vector--db-qdrant-informational.svg)](infrastructure/kubernetes/phase-5.0/vector-db/)
+[![Local LLM](https://img.shields.io/badge/local--llm-ollama-important.svg)](infrastructure/kubernetes/phase-5.0/local-llm/)
+[![MCP Protocol](https://img.shields.io/badge/MCP-protocol-blueviolet.svg)](infrastructure/kubernetes/phase-5.0/mcp/)
 
 > **🏆 World-Class Enterprise AI Research Platform - Production Ready**
 
-**The Free Deep Research System** is a complete, enterprise-grade AI-powered research platform that rivals industry leaders like Databricks, Snowflake, and Salesforce. Built with cloud-native architecture, advanced MLOps capabilities, multi-tenant support, and enterprise security compliance.
+**The Free Deep Research System** is a complete, enterprise-grade AI-powered research platform that rivals industry leaders like Databricks, Snowflake, and Salesforce. Built with cloud-native architecture, advanced MLOps capabilities, multi-tenant support, enterprise security compliance, and cutting-edge AI enhancements.
 
-**✨ From concept to enterprise-ready platform in 6 months** - featuring automated ML pipelines, real-time analytics, multi-tenant architecture, and zero-trust security.
+**✨ From concept to enterprise-ready platform in 7 months** - featuring automated ML pipelines, real-time analytics, multi-tenant architecture, zero-trust security, RAG capabilities, local LLM integration, and hybrid AI optimization.
 
-**Last Updated**: December 21, 2024
-**Status**: ✅ **PRODUCTION READY** - Ready for enterprise deployment
+**Last Updated**: July 21, 2025
+**Status**: ✅ **PRODUCTION READY** - Ready for enterprise deployment with Phase 5.0 AI Enhancement
 
 ## 🎯 **Enterprise Capabilities**
+
+### 🧠 **Phase 5.0: Advanced AI Enhancement (NEW)**
+- **RAG (Retrieval-Augmented Generation)**: Semantic search with vector embeddings and context retrieval
+- **Vector Database**: Qdrant v1.11.0 with high-performance vector storage and similarity search
+- **Local LLM Integration**: Ollama-powered local inference with GPU acceleration and model management
+- **Hybrid AI Optimization**: Intelligent model routing, cost optimization, and performance monitoring
+- **Multi-Provider Support**: OpenAI, Hugging Face, Groq, Together AI, Replicate integration
+- **MCP Protocol**: Model Context Protocol server for standardized AI model communication
+- **Enhanced BMAD Agents**: ML-specialized AI agents for RAG optimization and cost management
 
 ### 🤖 **Advanced AI/ML Operations (MLOps)**
 - **Kubeflow Pipelines**: Automated ML workflow orchestration
@@ -60,7 +73,7 @@
 git clone https://github.com/huggingfacer04/free-deep-research.git
 cd free-deep-research
 
-# Complete enterprise deployment
+# Complete enterprise deployment with Phase 5.0 AI Enhancement
 cd scripts
 ./production-startup.sh
 ```
@@ -73,6 +86,7 @@ cd infrastructure/kubernetes
 ./deploy-phase-4.7.sh  # Analytics
 ./deploy-phase-4.8.sh  # Enterprise
 ./deploy-phase-4.9.sh  # Security
+./deploy-phase-5.0.sh  # AI Enhancement (NEW)
 ```
 
 ### **Prerequisites**
@@ -85,7 +99,7 @@ cd infrastructure/kubernetes
 
 ```
 ┌─────────────────────────────────────────────────────────────────────────┐
-│                    Free Deep Research System v4.9                      │
+│                    Free Deep Research System v5.0                      │
 │                     Enterprise Production Architecture                  │
 ├─────────────────────────────────────────────────────────────────────────┤
 │  Phase 4.1-4.2: Event Sourcing + CQRS Foundation                      │
@@ -103,6 +117,8 @@ cd infrastructure/kubernetes
 │  Phase 4.8: Multi-tenant Enterprise (Keycloak + RBAC + Billing)        │
 ├─────────────────────────────────────────────────────────────────────────┤
 │  Phase 4.9: Security & Compliance (Vault + Velero + Falco)             │
+├─────────────────────────────────────────────────────────────────────────┤
+│  Phase 5.0: AI Enhancement (RAG + Vector DB + Local LLM + MCP)         │
 └─────────────────────────────────────────────────────────────────────────┘
 ```