… PHASE_5_0_IMPLEMENTATION_REPORT.md
… TASK_STATUS.md
… bmad-agent/agent-config.txt
… infrastructure/kubernetes/phase-5.0/ai-providers/provider-gateway.yaml
… infrastructure/kubernetes/phase-5.0/graphql/enhanced-schema.yaml
… infrastructure/kubernetes/phase-5.0/local-llm/gpu-resources.yaml
… infrastructure/kubernetes/phase-5.0/local-llm/ollama-deployment.yaml
… infrastructure/kubernetes/phase-5.0/mcp/mcp-server.yaml
… infrastructure/kubernetes/phase-5.0/optimization/cost-optimizer.yaml
… infrastructure/kubernetes/phase-5.0/optimization/model-router.yaml
… infrastructure/kubernetes/phase-5.0/rag/document-processor.yaml
… infrastructure/kubernetes/phase-5.0/rag/embedding-service.yaml
… infrastructure/kubernetes/phase-5.0/rag/rag-service.yaml
… infrastructure/kubernetes/phase-5.0/vector-db/qdrant-deployment.yaml
… infrastructure/kubernetes/phase-5.0/vector-db/qdrant-pvc.yaml
… package.json
… packages/mcp-server/package.json
… packages/rag-engine/package.json